Add initial version of ServiceGraph CRD

+++ go/apis/cluster/v1/networklinkqos.go
@@ -39,13 +39,13 @@ type NetworkLatency struct {
 	//
 	// +kubebuilder:validation:Minimum=0
 	// +kubebuilder:default=0
-	PacketDelayMsec int64 `json:"packetDelayMsec"`
+	PacketDelayMsec int32 `json:"packetDelayMsec"`
 
 	// The variance of PacketDelayMsec (i.e., jitter).
 	//
 	// +kubebuilder:validation:Minimum=0
 	// +kubebuilder:default=0
-	PacketDelayVariance int64 `json:"packetDelayVariance"`
+	PacketDelayVariance int32 `json:"packetDelayVariance"`
 }
 
 // NetworkPacketLoss describes the packet loss of a NetworkLink.
@@ -59,5 +59,5 @@ type NetworkPacketLoss struct {
 	// +kubebuilder:validation:Minimum=0
 	// +kubebuilder:validation:Maximum=10000
 	// +kubebuilder:default=0
-	PacketLossBp int64 `json:"packetLossBp"`
+	PacketLossBp int32 `json:"packetLossBp"`
 }
